python/scrape/requirements.txt
anyio==4.6.2.post1
APScheduler==3.10.4
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
pyzmq==26.2.0
selenium==4.26.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
